requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
llama-index-llms-google-genai
llama-index-embeddings-google-genai